# docs/openapi.yaml
openapi: 3.0.3
info:
  title: aniani api
  description: api for aniani reflectivity information
  contact:
    email: adelagarza@keck.hawaii.edu
  version: 1.0.0
servers:
  - url: http://vm-dev-appserver:5000/
paths:
  /getAllSamples:
    get:
      description: Returns all samples from a speciic mirror, telescope and measurement type. Every row in the MirrorSamples table that matches those will be returned.
      parameters:
        - name: mirror
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/mirror'
        - name: telescope_num
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/telescope_num'
        - name: measurement_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/measurement_type'
      responses:
        '200':
          description: Returns a list of dictionaries with each dict being a specific row in the table.
          content:
            json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/get_all_samples_output'
                
                
      
      
      
  /addReflectivityMeasurement:
    post:
      description: Add new samples into the MirrorSamples table.
      parameters:
      
    
    
    
    
    
  /getCurrentReflectivity:
    get:
      description: Returns a list of dictionaries for all the data of the active segment(s) (based on most recent install date) that are currently on for a given telescope, measurement_type, and mirror
      parameters:
        - name: mirror
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/mirror'
        - name: telescope_num
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/telescope_num'
        - name: measurement_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/measurement_type'
      responses:
        '200':
          description: Returns a list of dictionaries where each dictionary is the current reflectivity information of the segment / mirror on the telescope.
          content:
            json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/get_current_reflectiviy_output'
                  
                  
                  
                  
                  
  /getPredictReflectivity:
    get:
      description: Returns a list of dictionaries for all the data of the active segment(s) (based on most recent install date) that are currently on for a given telescope, measurement_type, and mirror
      parameters:
        - name: mirror
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/mirror'
        - name: telescope_num
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/telescope_num'
        - name: measurement_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/reflectivity_input_schema/properties/measurement_type'
            
      responses:
        '200':
          description: returns 4 dictionaries, one for each wavelength with the corresponding mean values and segment values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/get_predict_reflectiviy_output'
      
  
components:
  schemas:
    reflectivity_input_schema:
      type: object
      properties:
        mirror: 
          type: string
          enum:
            - 'primary'
            - 'secondary'
            - 'tertiary'
        telescope_num:
          type: integer
          enum:
            - 1
            - 2
        measurement_type:
          type: string
          enum:
            - 'T'
            - 'D'
            - 'S'
      required:
        - mirror
        - telescope_num
        - measurement_type
        
    add_reflectivity_measurement_schema:
      type: object
      properties:
        mirror: 
          type: string
          enum:
            - 'primary'
            - 'secondary'
            - 'tertiary'
        segment_id:
        mirror_type:
        measured_date:
        install_date:
        sample_status:
        telescope_num:
        segment_position:
        spectrum:
        measurement_type:
        reflectivity:
        notes:
        
        
    get_all_samples_output:
      type: object
      properties:
        id:
          type: number
          example: 4643
        install_date:
          type: string
          example: '2022-10-27'
        is_deleted:
          type: integer
          example: 0
        measured_date:
          type: string
          example: '2022-10-27'
        mirror:
          type: string
          example: primary
        measurement_type:
          type: string
          example: T
        mirror_type:
          type: string
          example: '1'
        segment_id:
          type: integer
          example: 49     
        segment_position:
          type: integer
          example: 18
        spectrum:
          type: string
          example: '400-540'
        telescope_num:
          type: integer
          example: 1
        notes:
          type: string
          example: 'testing'
        reflectivity:
          type: number
          example: 0.91
        sample_status:
          type: string
          example: 'clean'


    get_current_reflectiviy_output:
      type: object
      properties:
        400-540:
          type: number
          example: 0.909
        480-600:
          type: number
          example: 0.906
        590-720:
          type: number
          example: 0.895
        900-1100:
          type: number
          example: 0.891
        install_date:
          type: string
          example: Wed, 28 April 2021
        measured_date:
          type: string
          example: Sat, 16 Jan 2021
        mirror:
          type: string
          example: primary
        measurement_type:
          type: string
          example: T
        mirror_type:
          type: string
          example: 1
        segment_id:
          type: integer
          example: 49
          
          
    get_predict_reflectiviy_output:
      type: object
      properties:
        spectrum:
          type: object
          properties:
            dR/year:
              type: number
              example: -0.0015166
            r_avg:
              type: number
              example: 0.915104
            segments:
              type: object
              properties:
                segment_position:
                  type: object
                  properties:
                    install_date:
                      type: string
                      example: Wed, 28 April 2021
                    measured_date:
                      type: string
                      example: Sat, 16 Jan 2021
                    mirror:
                      type: string
                      example: primary
                    measurement_type:
                      type: string
                      example: T
                    mirror_type:
                      type: string
                      example: 1
                    segment_id:
                      type: integer
                      example: 49
                    predict_reflectivity:
                      type: number
                      example: 0.92
      
          